fix some lua_tolstring issues (And the previous issues with createStringstream failing even when using size_t still failing i'll chalk up to compiler caches)

+++ Cheat Engine/luapipe.pas
@@ -644,7 +644,7 @@ var
   p: TPipeconnection;
   paramcount: integer;
   s: pchar;
-  slength: integer;
+  slength: size_t;
   include0terminator: boolean;
 begin
   result:=0;
